!fix: Only inherit sensible attribute from parents
If you are using the `private` key for tags, you will need to
switch to using `tags`, or the appropriate top level tag attribute
like `sql_query` or `http`.

This is also 1 considered a breaking change because it is
likely/possible that implementations were relying on this
behavior. If you aren't leveraging tags, or if your
spans aren't very deep in general than you most likely
don't need to deal with this.