allow headers to be passed into `Spandex.distributed_context/2`
This change allows headers to be specified in
`Spandex.distributed_context/2` as an alternative to `Plug.Conn`. This
is more compatible with the current signature of
`Spandex.inject_context/3` and allows Spandex to be used to extract
trace information from connections that may not be relying on Plug e.g.
GRPC.

Fixes #112